+++ tests/TestCase.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Nuewire\Installer\Tests;
 
+use Livewire\LivewireServiceProvider;
 use Nuewire\Installer\InstallerServiceProvider;
 use Orchestra\Testbench\TestCase as Orchestra;
 
@@ -11,6 +12,6 @@ abstract class TestCase extends Orchestra
 {
     protected function getPackageProviders($app): array
     {
-        return [InstallerServiceProvider::class];
+        return [LivewireServiceProvider::class, InstallerServiceProvider::class];
     }
 }